Set up full backend routes

+++ routes/user.js
@@ -2,6 +2,7 @@ const express = require("express");
 const router = express.Router();
 const creatError = require("http-errors");
 const User = require("../models/user");
+const Project = require("../models/project");
 const mongoose = require('mongoose');
 
 // HELPER FUNCTIONS
@@ -11,20 +12,32 @@ const {
     isNotLoggedIn
 } = require("../helpers/middlewares");
 
+// GET '/user/all'
+router.get('/all', isLoggedIn,  async (req,res,next) => {
+    try {
+        const user = await User.find()
+            .populate('initiatorOnProject appliedOnProject acceptedOnProject');
+        res.status(200).json(user);
+    }
+    catch (error) {
+        next(error);
+    }
+})
+
 // GET '/user/:id'
 //isLoggedIn,
 
-router.get('/:id',   async (req,res,next) => {
+router.get('/:id', isLoggedIn,  async (req,res,next) => {
     try {
         const { id } =req.params;
 
         if ( !mongoose.Types.ObjectId.isValid(id)) {
-            res.status(400).json({message: 'Unvalid Id'});
+            res.status(400).json({"message": "Unvalid Id"});
             return;
         }
 
-        const user = await User.findById(id);
-            // .populate('');
+        const user = await User.findById(id)
+            .populate('initiatorOnProject appliedOnProject acceptedOnProject');
         res.status(200).json(user);
     }
     catch (error) {
@@ -71,24 +84,25 @@ router.put('/edit/:id', isLoggedIn, async (req, res, next) => {
 // DELETE '/user/:id'
 router.delete('/:id', isLoggedIn, async(req, res, next) => {
     
+    const { id } = req.params;
+    
     // check if user trying to edit is the loggedin user
-    // if (id != req.session.currentUser._id) {
-    //     res.status(401).json({message: 'Unauthorized user id'});
-    //     return;
-    // } 
+    if (id != req.session.currentUser._id) {
+        res.status(401).json({message: 'Unauthorized user id'});
+        return;
+    } 
 
     try {
-        const { id } = req.params;
         
         await User.findByIdAndRemove(id);
-
+        req.session.destroy();
         // await Project.updateMany(
         //     {},
         //     {$pull: { appliedParticipants:id, acceptedParticipants:id}},
         //     { new: true }
         // );
 
-        res.status(200).json({message:'User deleted'})
+        res.status(200).json({"message":"User deleted"})
     }
     catch(error) {
         next (error);
